refactor(refactoring main files): changing and destructuring main files for refactoring
