+++ src/config/gameOptions.ts
@@ -1,16 +1,16 @@
 // Opções de configuração
-import { PreloadAssets } from "../scenes/preloadAssets";
+import { PreloadAssets } from '../assets/preloadAssets';
 
 export const GameOptions = {
   gameSize: {
     width: 1920,
     height: 1080,
   },
-    gameBackground: 'gameBackgroundLimbo',
-    type: Phaser.AUTO,
-    playerSpeed: 100, // velocidade do jogador em pixels por segundo
-    enemySpeed: 50, // velocidade do inimigo, em pixels por segundo
-    bulletSpeed: 100, // velocidade da bala, em pixels por segundo
-    bulletRate: 1500, // rate da bala, balas por milisegundos
-    enemyRate: 800, // rate do inimigo, inimigos por milisegundos
+  gameBackground: 'gameBackgroundLimbo',
+  type: Phaser.AUTO,
+  playerSpeed: 100, // velocidade do jogador em pixels por segundo
+  enemySpeed: 50, // velocidade do inimigo, em pixels por segundo
+  bulletSpeed: 100, // velocidade da bala, em pixels por segundo
+  bulletRate: 1500, // rate da bala, balas por milisegundos
+  enemyRate: 800, // rate do inimigo, inimigos por milisegundos
 };
